+++ src/main/java/city/Drone.java
@@ -1,5 +1,10 @@
 package city;
 
+import filestuff.FileIO;
+import users.Tourists;
+import users.User;
+
+import java.io.File;
 import java.util.Scanner;
 
 public class Drone
@@ -36,6 +41,30 @@ public class Drone
     }
 
     //end of set and get functions
+
+    public void AddUser(Tourists user,String filename)
+    {
+        File file = new File("/home/cybereagle3-1/IdeaProjects/midsem-project-csd-207/src/main/java/filestuff/" + filename);
+        new FileIO().WriteUsersonFile(user, file);
+    }
+
+    public void WhereAmI(User user)
+    {
+        System.out.println("X=" + user.GetUserCoordsX());
+        System.out.println("Y=" + user.GetUserCoordsY());
+    }
+
+    public void MoveDrone(User user)
+    {
+        Scanner input = new Scanner(System.in);
+
+        System.out.println("enter coordinates");
+        System.out.print("X:");
+        int x = Integer.parseInt(input.nextLine());
+        System.out.print("Y:");
+        int y = Integer.parseInt(input.nextLine());
+        user.SetUserCoords(x,y);
+    }
 }