+++ src/main/java/city/TouristDrone.java
@@ -39,7 +39,7 @@ public class TouristDrone extends Drone
 
             Places landmark = new Places();
 
-            System.out.println("enter the landmark details.\n name:");
+            System.out.print("enter the landmark details.\n name:");
             landmark.SetName(input);
             System.out.println("picture ID");
             landmark.SetPictureID(input);
@@ -56,30 +56,6 @@ public class TouristDrone extends Drone
         }
 
         return;
-    }
-
-    public void AddTourist(Tourists tourist)
-    {
-        File file = new File("/home/cybereagle3-1/IdeaProjects/midsem-project-csd-207/src/main/java/filestuff/tourists.txt");
-        new FileIO().WriteUsersonFile(tourist, file);
-    }
-
-    public void WhereAmI(Tourists tourist)
-    {
-        System.out.println("X=" + tourist.GetUserCoordsX());
-        System.out.println("Y=" + tourist.GetUserCoordsY());
-    }
-
-    public void MoveDrone(Tourists tourist)
-    {
-        Scanner input = new Scanner(System.in);
-
-        System.out.println("enter coordinates");
-        System.out.print("X:");
-        int x = Integer.parseInt(input.nextLine());
-        System.out.print("Y:");
-        int y = Integer.parseInt(input.nextLine());
-        tourist.SetUserCoords(x,y);
     }
 
 }